♻️ refactor: 프로필 헤더 컴포넌트로 분리
